refactor: move due date state logic to a hook, to be reusable
